fix(vision): move inference rescale helpers to a PIL-free module
Importing inference_scale/scale_box/scale_point from ai.common.image.dense_resize
ran the image package __init__ -> PIL, which broke face_detection's PIL-free
rescale unit test on CI. Move the three pure helpers to ai.common.utils.image_utils
— lazy/PIL-free, and already the shared image-helper home.
