+++ packages/ai/src/ai/common/image/dense_resize.py
@@ -4,7 +4,7 @@ See vision suite build brief §1.3 and Plan #13. Primary strategy is max-edge re
 """
 
 from __future__ import annotations
-from typing import Dict, Literal, Optional, Tuple, Union
+from typing import Literal, Optional, Tuple, Union
 import numpy as np
 from PIL import Image
 
@@ -69,41 +69,6 @@ def resize_for_inference(
 
     resized = image.resize((new_w, new_h), resample=Image.LANCZOS)
     return resized, (orig_w, orig_h)
-
-
-def inference_scale(small_size: Tuple[int, int], original_size: Tuple[int, int]) -> Optional[Tuple[float, float]]:
-    """(fx, fy) mapping coords from a downscaled inference image back to the original.
-
-    The sparse-output counterpart to ``restore_dense_output``: detection / pose / face
-    run inference on a ``resize_for_inference`` downscale, then scale their box,
-    keypoint and centroid coords by these factors.
-
-    Args:
-        small_size: (width, height) the inference ran on.
-        original_size: (width, height) of the original image.
-
-    Returns:
-        (fx, fy) scale factors, or None when the sizes already match (no rescale needed).
-    """
-    sw, sh = int(small_size[0]), int(small_size[1])
-    ow, oh = int(original_size[0]), int(original_size[1])
-    if sw == ow and sh == oh:
-        return None
-    return ow / sw, oh / sh
-
-
-def scale_box(box: Dict[str, float], fx: float, fy: float) -> None:
-    """Scale an ``{x1, y1, x2, y2}`` box in place by (fx, fy)."""
-    box['x1'] *= fx
-    box['x2'] *= fx
-    box['y1'] *= fy
-    box['y2'] *= fy
-
-
-def scale_point(point: Dict[str, float], fx: float, fy: float) -> None:
-    """Scale an ``{x, y}`` point (keypoint / centroid / landmark) in place by (fx, fy)."""
-    point['x'] *= fx
-    point['y'] *= fy
 
 
 def restore_dense_output(
